Some cleanups
Remove rtkLoader.js -- it was for loading rtk.csv, which is no longer used.

loaderHelper.js:
Move otherDownloads out of determineNumTasks, since those are not counted,
and I want to remodularize the ones that are counted, a lot.

loaderUtilities.js:
Add some things that are not yet used.
Move applyRotation here, from rtkLoader.js.

rtkLoaderFlatbuffer.js:
Move animateRTK here, from rtkLoader.js.
