+++ demo/rtkLoaderFlatbuffer.js
@@ -1,11 +1,10 @@
 'use strict';
 import { incrementLoadingBarTotal, updateLoadingBar } from "../common/overlay.js";
 import { RtkTrajectory } from "../demo/RtkTrajectory.js";
-import { animateRTK } from "../demo/rtkLoader.js";
-import { } from  "../demo/paramLoader.js";
 import { loadTexturedCar } from "../demo/textureLoader.js";
 import { getFbFileInfo } from "./loaderUtilities.js";
-
+import { visualizationMode } from "../demo/paramLoader.js";
+import { applyRotation } from "../demo/loaderUtilities.js";
 
 let rtkFiles = null;
 // sets local variable and returns so # files can be counted
@@ -20,7 +19,7 @@ export const rtkFlatbufferDownloads = async (datasetFiles) => {
 
 export async function loadRtkFlatbuffer(s3, bucket, name, callback) {
   if (!rtkFiles) {
-    console.log("No flatbuffer files present")
+    console.log("No rtk files present")
     return
   }
 
@@ -182,5 +181,76 @@ export function loadRtkCallback(s3, bucket, name, callback) {
 
 		// RTK TweenTarget Callback:
 		animateRTK();
+	});
+}
+
+
+ // animates the viewer camera and TweenTarget for RTK
+// once textured vehicle object is created
+function animateRTK() {
+	window.updateCamera = true;
+	window.pitchThreshold = 1.00;
+	animationEngine.tweenTargets.push((gpsTime) => {
+		try {
+			let t = (gpsTime - animationEngine.tstart) / (animationEngine.timeRange);
+			// vehicle contains all the work done by textureLoader
+			let vehicle = viewer.scene.scene.getObjectByName("Vehicle");
+                        const mesh = vehicle.getObjectByName("Vehicle Mesh");
+                        const meshPosition = new THREE.Vector3();
+		        let lastRtkPoint = vehicle.position.clone();
+			let lastRtkOrientation = vehicle.rotation.clone();
+			let lastTransform = vehicle.matrixWorld.clone();
+			// debugger; //vehicle
+			let state = vehicle.rtkTrajectory.getState(gpsTime);
+			let rtkPoint = state.pose.clone();
+			let vehicleOrientation = state.orient.clone();
+			vehicle.position.copy(rtkPoint);
+			if (visualizationMode == "aptivLanes") {
+				vehicle.position.add(new THREE.Vector3(0, 0, 1000));
+			}
+			applyRotation(vehicle, vehicleOrientation.x, vehicleOrientation.y, vehicleOrientation.z);
+			vehicle.updateMatrixWorld();
+
+			// Apply Transformation to Camera and Target:
+			if (window.updateCamera) {
+				let newTransform = vehicle.matrixWorld.clone();
+				let lastTransformInverse = lastTransform.getInverse(lastTransform);
+				let deltaTransform = lastTransformInverse.premultiply(newTransform);
+				let target = viewer.scene.view.position.clone();
+				let direction = viewer.scene.view.direction.clone();
+				let radius = viewer.scene.view.radius;
+				target.add(direction.multiplyScalar(radius));
+				viewer.scene.view.position.applyMatrix4(deltaTransform);
+				if (Math.abs(viewer.scene.view.pitch) < window.pitchThreshold) {
+					viewer.scene.view.lookAt(target.applyMatrix4(deltaTransform));
+				}
+			}
+
+			// Set Elevation:
+			// let elevationDeltaMin = -0;
+			// let elevationDeltaMax = 2;
+			let clouds = viewer.scene.pointclouds;
+                        const elevationWindow = window.animationEngine.elevationWindow;
+                        const {min, max} = elevationWindow;
+                        const elevationMin = Number(min);
+                        const elevationMax = Number(max);
+			for (let ii = 0, numClouds = clouds.length; ii < numClouds; ii++) {
+                                meshPosition.setFromMatrixPosition(mesh.matrixWorld);
+                                const zheight = meshPosition.z;
+                                // Is this setting ever used?
+			        elevationWindow.z = zheight;
+			        viewer.scene.pointclouds[ii].material.elevationRange = [zheight + elevationMin, zheight + elevationMax];
+				// TODO set elevation slider range extent
+			}
+
+			// Save Current RTK Pose in Uniforms:
+			for (let ii = 0, numClouds = clouds.length; ii < numClouds; ii++) {
+				let material = clouds[ii].material;
+				material.uniforms.currentRtkPosition.value = state.pose.clone();
+				material.uniforms.currentRtkOrientation.value = state.orient.clone();
+			}
+		} catch (e) {
+			console.error("Caught error: ", e);
+		}
 	});
 }
